requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
dotenv==0.9.9
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.13.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.38.2
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.3.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.45.0
supervision==0.25.1
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.128
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
